Add TOML support
Add support for handling ID3 metadata presented as TOML. This can be
useful for building automation around id3manager (TOML is machine
parseable, unlike plain text).

A new command line argument is added to allow selection of the metadata
format (--format, or -f). The plain text format is the default.

We use this opportunity to turn id3manager into a package with
multiple subpackages and modules. Entrypoints and CLI invocations
remain unchanged.

Co-authored-by: Roman Podoliaka <roman.podoliaka@gmail.com>